Add new functionality and try to move calculations to non-GUI class
I added a wind vector calculation if we are receiving magnetic heading,
airspeed, and GPS data. I put the calculation into a class that doesn't
draw anything and then update a widget with the calculated values. Maybe
this will be helpful for performance reasons at some point. Almost
finsihes #3, I just need to work on the display which will just be two
arrows that will reuse the existing wind vector drawing.