Wind Component display
Fixes #3

The user can decide to view the relative wind as a vector or in its X
and Y components. May still need to do some work to make it prettier,
but the functionality is there. Also will need to add a way for the user
to select the display.